refactor: Move multer and cloudinary logic into `/utils`
